requirements.txt
fastapi

uvicorn[standard]

psycopg2-binary

sqlalchemy

python-dotenv

pydantic

fastapi[all]

loguru